Double Clicking Support for Spells (#8)
* Update React-DOM types, add createRoot

* React 18 compat

* Scov consideration for SW:D

* Separate spell icons

* Double click support

* Quick cleanup of imports
