+++ package.json
@@ -13,7 +13,7 @@
     "@types/node": "^17.0.2",
     "@types/react": "^17.0.0",
     "@types/react-copy-to-clipboard": "^5.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react-dom": "^18.0.6",
     "@types/styled-components": "^5.1.10",
     "@types/uuid": "^8.3.3",
     "classnames": "^2.3.2",
