+++ src/index.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
+
 import "./index.css";
 import "tippy.js/dist/tippy.css";
 import App from "./App";
@@ -9,7 +10,14 @@ import { HTML5Backend } from "react-dnd-html5-backend";
 import { SimulationsProvider } from "context/simulations";
 import { QuickFill } from "features/QuickFill";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+if (container === null) {
+  throw new Error("Could not find root!");
+}
+
+const root = createRoot(container);
+
+root.render(
   <React.StrictMode>
     <SimulationsProvider>
       <DndProvider backend={HTML5Backend}>
@@ -17,8 +25,7 @@ ReactDOM.render(
         <App />
       </DndProvider>
     </SimulationsProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
